poocs-web/backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
websockets>=12.0
openai>=1.3.5
pydantic>=2.5.2
pytest>=7.4.3
httpx>=0.25.2  # For testing
python-multipart>=0.0.6  # For handling form data
